requirements.txt
pyannote.audio
resemblyzer
speechbrain
torchaudio
torch
soundfile
librosa
numpy
scikit-learn
tqdm
openai-whisper
python-dotenv
pydub
jiwer